Update README.md to correctly include screenshots for this lab that were updated into the Assignment 1 repository by accident and mistake.
The screenshots for this lab had been uploaded on time, but had been mistakenly edited into the README.md of the Assignment 1 repository at https://github.com/AndyXi163/ECE444-F2023-Assignment1.git .
Please double-check the commits in that repository for references.
This commit is simply for the purpose of clarification and correction. I will make necessary changes and reverts for both repositories after grading is complete.